+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## unreleased
+### Fixed
+ - bug in lexer that was capturing the carriage return character (`\n`) at the end of comment statements
+
 
 ## [0.9.4] - 2020-05-05
 ### Added